backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
azure-identity
python-dotenv
httpx
fpdf
azure-storage-blob
python-jose
pyyaml
